refactor(capture-sdk): Add helper method to `FailureException` for handling casting from `Throwable`
Also convert `FailureException´ to kotlin and move it the `internal.network` package.

PIA-3628
